requirements.txt
torch
torchvision
gradio
numpy
Pillow
